requirements.txt
ultralytics
opencv-python
open3d
numpy
matplotlib
Pillow
pytest
PyYAML
